pyproject.toml
[project]
name = "claude-rest-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cryptography>=46.0.6",
    "django>=6.0.3",
    "django-allauth[socialaccount]>=65.15.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "psycopg>=3.3.3",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.8",
]
